Fix bug in tags - don’t include participating addresses multiple times
UInt8Arrays are classes, if I create a new one it’s not going to work in a set
one day i will be good at typescript

test plan
Updated the tags.test to be a unit test since it doesn’t do any network calls
refactored out the duplicate code creating channel messages
verified that the test still passed
added a second message in the thread
verified that the test failed
fixed the code
